fix(sidebar): integrate upstream Codex subagent models
