+++ src/renderer/src/components/sidebar/worktree-card-compact-agent-row.subagent.test.tsx
@@ -53,7 +53,7 @@ describe('CompactAgentRow subagent display', () => {
     expect(row.textContent).toContain('test_agent')
     expect(row.textContent).not.toContain('/root/')
     expect(row.textContent).not.toContain('Default')
-    expect(row.textContent).not.toContain('gpt-5.6-sol')
+    expect(row.textContent).toContain('gpt-5.6-sol')
     expect(row.getAttribute('title')).toBe('/root/test_agent')
     expect(row.getAttribute('data-subagent-id')).toBe('child-1')
     expect(row.getAttribute('data-current')).toBe('true')
